requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==11.3.0
PyJWT==2.10.1
reportlab==4.4.5
requests==2.32.5
sqlparse==0.5.3
stripe==14.0.1
typing_extensions==4.15.0
urllib3==2.6.0
boto3==1.34.0
django-storages==1.14.2
python-dotenv==1.2.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
dj-database-url==2.1.0
